requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
ultralytics==8.3.87
torch==2.6.0
opencv-python==4.11.0.86
Pillow==11.1.0
numpy==2.1.1
fpdf==1.7.2
gunicorn==21.2.0